requirements.txt
fastapi
uvicorn
openai
pinecone
python-jose
python-multipart
PyJWT
python-dotenv
